requirements.txt
Flask==3.0.2
google-cloud-firestore==2.14.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
waitress==3.0.0
requests==2.31.0
